Simplify touch_routine by using native int instead

+++ ESP32-fluid-simulation/ESP32-fluid-simulation.ino
@@ -67,7 +67,7 @@ void touch_routine(void *args){
   ts_spi.begin(XPT2046_CLK, XPT2046_MISO, XPT2046_MOSI, XPT2046_CS);
   ts.begin(ts_spi);
 
-  Vector2<uint16_t> last_coords, current_coords; // used for calculating the velocity
+  Vector2<int> last_coords, current_coords; // used for calculating the velocity
   bool last_touched = false, touched; // used for detecting when to send a touch struct
 
   while(1){
@@ -80,9 +80,9 @@ void touch_routine(void *args){
 
       // We need to map from a 4096x4096 domain to a N_ROWSxN_COLS one
       TS_Point raw_coords = ts.getPoint();
-      current_coords = (Vector2<uint16_t>){
-          .x = (uint16_t)(raw_coords.x * N_COLS / 4096),
-          .y = (uint16_t)(raw_coords.y * N_ROWS / 4096)};
+      current_coords = (Vector2<int>){
+          .x = raw_coords.x * N_COLS / 4096,
+          .y = raw_coords.y * N_ROWS / 4096};
     }
     // else current_coords should never end up being used
 
@@ -95,8 +95,8 @@ void touch_routine(void *args){
     if(send_touch){
       // calculate and send the velocity and location
       Vector2<float> current_velocity = {
-          .x = ((float)current_coords.x - (float)last_coords.x) * 1000 / POLLING_PERIOD,
-          .y = ((float)current_coords.y - (float)last_coords.y) * 1000 / POLLING_PERIOD};
+          .x = (current_coords.x - last_coords.x) * 1000.f / POLLING_PERIOD,
+          .y = (current_coords.y - last_coords.y) * 1000.f / POLLING_PERIOD};
       struct touch current_touch = { .coords = current_coords, .velocity = current_velocity };
       xQueueSend(touch_queue, &current_touch, 0); // TODO: don't just use send and pray
     }
